requirements.txt
unstructured
langchain
openai
pdf2image
chromadb
tiktoken
python-dotenv
pypdf
streamlit